Do not attempt digest re-authentication
Firefox does not display the auth dialog and fail all subsequent auth
attempts following a canceled one. This results in a endless flow of
failed auth errors. In case of just a wrong password the broser
re-opens the dialog on its own allowing the user to correct the
login name or password.

Fixes: SL1SW-1360