Wrap filename with 'anywhere' behavior
Rather than having the filename run on for one line even when it is
particularly long (as is commonly the case with gcode filenames). Have
the filename wrap into another line. This works on both desktop and
mobile layouts with a big improvement to the look of the webpage on
mobile.